Change GalleryToBootstrapFrom to Repository
Just choosing to use the parameter name that matches what this thing is (in PowerShell it's a `Repository` or a `PSRepository` or a `PackageSource`).  I thought about using "ToolsPackageSource" but since this build script doesn't currently use repositories for anything else, it seems like we might as well just keep it simple.